S5收官: 审计五线索清零(GetTickCount真冲突LP64统一DWORD/f53c8重复移植去冗/b3l死条件按dc修复/b1过时注释更新/a584正式UNIMPLEMENTED登记)

+++ link_stubs.c
@@ -107,7 +107,7 @@ DWORD   GetCurrentThreadId(void);
 DWORD   GetWindowThreadProcessId(HWND w, DWORD *pid);
 void   *LoadLibraryA(const char *m);
 FARPROC GetProcAddress(void *hm, const char *name);
-unsigned long GetTickCount(void);
+DWORD   GetTickCount(void);   /* 统一签名 (原 unsigned long 与 DWORD 定义冲突, LP64 下 64位) */
 uint64_t GetProcessHeap(void);
 int64_t *PECMD_AssignString(int64_t *, const uint16_t *);
 void *TlsGetValue(unsigned long idx);
